# Eliminate bounds check in interpreter hot path