@Robbepop 2022-01-07
> I can't give you absolution for this but I think your proof checks out. The validator should statically prevent jumping to an invalid address.

One thing I want to do before we merge this is to reintroduce the bounds check for debug builds with debug_assert. so we can still safely test wasmi without optimization in case we are changing affected code parts.